@@ -61,7 +61,7 @@
     "@commitlint/cli": "^19.0.0",
     "@commitlint/config-angular": "^19.0.0",
     "@ctrl/tinycolor": "^3.6.0",
-    "@eslint/js": "^9.34.0",
+    "@eslint/js": "^9.36.0",
     "@eslint/markdown": "^7.0.0",
     "@inquirer/prompts": "^7.0.0",
     "@schematics/angular": "^20.2.0",
@@ -77,8 +77,9 @@
     "@types/minimist": "^1.2.5",
     "@types/node": "^20.11.1",
     "@types/pascalcase": "^1.0.3",
+    "@types/prismjs": "^1.26.5",
     "@types/yaml-front-matter": "^4.1.3",
-    "angular-eslint": "^20.2.0",
+    "angular-eslint": "^20.3.0",
     "chalk": "^4.1.0",
     "codecov": "^3.8.1",
     "conventional-changelog": "^7.1.1",
@@ -90,10 +91,10 @@
     "dagre-compound": "^0.0.8",
     "del": "^8.0.0",
     "detect-port": "^2.1.0",
-    "eslint": "^9.34.0",
+    "eslint": "^9.36.0",
     "eslint-plugin-header": "~3.1.1",
     "eslint-plugin-import": "^2.32.0",
-    "eslint-plugin-jsdoc": "^54.0.0",
+    "eslint-plugin-jsdoc": "^60.0.0",
     "eslint-plugin-prefer-arrow": "~1.2.2",
     "eslint-plugin-prettier": "~5.5.0",
     "eslint-plugin-unused-imports": "^4.1.4",
@@ -103,8 +104,8 @@
     "gulp-clean": "^0.4.0",
     "html-minifier-terser": "^7.2.0",
     "husky": "^9.0.0",
-    "jasmine": "^5.8.0",
-    "jasmine-core": "^5.8.0",
+    "jasmine": "^5.10.0",
+    "jasmine-core": "^5.10.0",
     "jasmine-spec-reporter": "^7.0.0",
     "karma": "^6.4.4",
     "karma-chrome-launcher": "^3.1.0",
@@ -131,13 +132,13 @@
     "prettier": "^3.6.0",
     "prismjs": "^1.29.0",
     "process": "^0.11.10",
-    "puppeteer": "^24.14.0",
+    "puppeteer": "^24.22.0",
     "remark": "^15.0.0",
     "resolve-bin": "^1.0.1",
     "rxjs": "~7.8.1",
     "sitemap": "^8.0.0",
     "slugify": "^1.6.6",
-    "stylelint": "^16.21.0",
+    "stylelint": "^16.24.0",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-config-standard": "^39.0.0",
     "stylelint-config-standard-less": "^3.0.1",
@@ -148,7 +149,7 @@
     "tslib": "^2.0.3",
     "tsx": "^4.20.3",
     "typescript": "~5.9.0",
-    "typescript-eslint": "^8.37.0",
+    "typescript-eslint": "^8.44.0",
     "yaml-front-matter": "^4.1.1",
     "zone.js": "~0.15.0"
   }
