@@ -12,21 +12,24 @@ import { Observable } from 'rxjs';
  * where streams of data are already being used and we need to wait until
  * the change detection ends before performing any tasks.
  */
-export const NZ_AFTER_NEXT_RENDER$ = new InjectionToken<Observable<void>>('nz-after-next-render', {
-  providedIn: 'root',
-  factory: () => {
-    const injector = inject(Injector);
+export const NZ_AFTER_NEXT_RENDER$ = new InjectionToken<Observable<void>>(
+  typeof ngDevMode !== 'undefined' && ngDevMode ? 'nz-after-next-render' : '',
+  {
+    providedIn: 'root',
+    factory: () => {
+      const injector = inject(Injector);
 
-    return new Observable<void>(subscriber => {
-      const ref = afterNextRender(
-        () => {
-          subscriber.next();
-          subscriber.complete();
-        },
-        { injector }
-      );
+      return new Observable<void>(subscriber => {
+        const ref = afterNextRender(
+          () => {
+            subscriber.next();
+            subscriber.complete();
+          },
+          { injector }
+        );
 
-      return () => ref.destroy();
-    });
+        return () => ref.destroy();
+      });
+    }
   }
-});
+);
