@@ -36,6 +36,7 @@ describe('Space compact', () => {
   it('should render all child components', () => {
     const spaceCompactElement: HTMLElement = fixture.nativeElement;
     const nzInput = spaceCompactElement.querySelector('input[nz-input]');
+    const nzInputWrapper = spaceCompactElement.querySelector('nz-input-wrapper');
     const nzInputGroup = spaceCompactElement.querySelector('nz-input-group');
     const nzInputNumber = spaceCompactElement.querySelector('nz-input-number');
     const nzDatePicker = spaceCompactElement.querySelector('nz-date-picker');
@@ -48,6 +49,7 @@ describe('Space compact', () => {
 
     expect(nzInput).toBeTruthy();
     expect(nzInputNumber).toBeTruthy();
+    expect(nzInputWrapper).toBeTruthy();
     expect(nzInputGroup).toBeTruthy();
     expect(nzDatePicker).toBeTruthy();
     expect(nzRangePicker).toBeTruthy();
@@ -58,6 +60,8 @@ describe('Space compact', () => {
     expect(nzButton).toBeTruthy();
 
     expect(nzInput!.classList).toContain('ant-input-compact-item');
+    expect(nzInputWrapper!.classList).toContain('ant-input-compact-item');
+    expect(nzInputWrapper!.querySelector('input.ant-input-compact-item')).toBeFalsy();
     expect(nzInputGroup!.classList).toContain('ant-input-compact-item');
 
     expect(nzInputNumber!.classList).toContain('ant-input-number-compact-item');
@@ -76,7 +80,7 @@ describe('Space compact', () => {
   it('should be possible to switch compact first / last classes', async () => {
     const spaceCompactElement: HTMLElement = fixture.nativeElement;
     const nzInput = spaceCompactElement.querySelector('input[nz-input]');
-    const nzInputGroup = spaceCompactElement.querySelector('nz-input-group');
+    const nzInputWrapper = spaceCompactElement.querySelector('nz-input-wrapper');
     const nzTreeSelect = spaceCompactElement.querySelector('nz-tree-select');
     const nzButton = spaceCompactElement.querySelector('button[nz-button]');
 
@@ -84,7 +88,7 @@ describe('Space compact', () => {
 
     expect(nzInput!.classList).toContain('ant-input-compact-first-item');
     expect(nzButton!.classList).toContain('ant-btn-compact-last-item');
-    expect(nzInputGroup!.classList).not.toContain('ant-input-compact-first-item');
+    expect(nzInputWrapper!.classList).not.toContain('ant-input-compact-first-item');
     expect(nzTreeSelect!.classList).not.toContain('ant-select-compact-last-item');
 
     component.showFirst = false;
@@ -93,13 +97,14 @@ describe('Space compact', () => {
 
     await Promise.resolve();
 
-    expect(nzInputGroup!.classList).toContain('ant-input-compact-first-item');
+    expect(nzInputWrapper!.classList).toContain('ant-input-compact-first-item');
     expect(nzTreeSelect!.classList).toContain('ant-select-compact-last-item');
   });
 
   it('should be apply size class', () => {
     const spaceCompactElement: HTMLElement = fixture.nativeElement;
     const nzInput = spaceCompactElement.querySelector('input[nz-input]');
+    const nzInputWrapper = spaceCompactElement.querySelector('nz-input-wrapper');
     const nzInputNumber = spaceCompactElement.querySelector('nz-input-number');
     const nzDatePicker = spaceCompactElement.querySelector('nz-date-picker');
     const nzRangePicker = spaceCompactElement.querySelector('nz-range-picker');
@@ -113,6 +118,9 @@ describe('Space compact', () => {
     fixture.detectChanges();
 
     expect(nzInput!.classList).toContain('ant-input-sm');
+    expect(nzInputWrapper!.classList).toContain('ant-input-group-wrapper-sm');
+    expect(nzInputWrapper!.querySelector('.ant-input-affix-wrapper-sm')).toBeTruthy();
+    expect(nzInputWrapper!.querySelector('input.ant-input-sm')).toBeTruthy();
     expect(nzInputNumber!.classList).toContain('ant-input-number-sm');
     expect(nzDatePicker!.classList).toContain('ant-picker-small');
     expect(nzRangePicker!.classList).toContain('ant-picker-small');
@@ -126,6 +134,9 @@ describe('Space compact', () => {
     fixture.detectChanges();
 
     expect(nzInput!.classList).toContain('ant-input-lg');
+    expect(nzInputWrapper!.classList).toContain('ant-input-group-wrapper-lg');
+    expect(nzInputWrapper!.querySelector('.ant-input-affix-wrapper-lg')).toBeTruthy();
+    expect(nzInputWrapper!.querySelector('input.ant-input-lg')).toBeTruthy();
     expect(nzInputNumber!.classList).toContain('ant-input-number-lg');
     expect(nzDatePicker!.classList).toContain('ant-picker-large');
     expect(nzRangePicker!.classList).toContain('ant-picker-large');
@@ -224,6 +235,11 @@ describe('Space compact direction', () => {
       @if (showFirst) {
         <input nz-input />
       }
+      <nz-input-wrapper>
+        <span nzInputAddonBefore>Before</span>
+        <span nzInputPrefix>Prefix</span>
+        <input nz-input />
+      </nz-input-wrapper>
       <nz-input-group><input nz-input /></nz-input-group>
       <nz-input-number />
       <nz-date-picker />
